Add recipe for preparing groundnut soup
Added a groundnut_soup.html file that contains a step-by-step procedure
for preparing groundnut soup, and linked this page in index.html.

I also refactored the code in index.html to improve readability.
Also included an introductory paragraph that talks about what the user
can find on the page and how the page can help the user to easily
prepare any of the dishes featured on the website.